feat: bulk actions support
