+++ README.md
@@ -101,6 +101,7 @@ There are two ways to run MediaCMS, through Docker Compose and through installin
 * [Configuration](docs/admins_docs.md#5-configuration) page
 * [Transcoding](docs/transcoding.md) page
 * [Developer Experience](docs/dev_exp.md) page
+* [Media Permissions](docs/media_permissions.md) page
 
 
 ## Technology
