+++ files/models/__init__.py
@@ -1,0 +1,25 @@
+# Import all models for backward compatibility
+from .category import Category, Tag  # noqa: F401
+from .comment import Comment  # noqa: F401
+from .encoding import EncodeProfile, Encoding  # noqa: F401
+from .license import License  # noqa: F401
+from .media import Media, MediaPermission  # noqa: F401
+from .playlist import Playlist, PlaylistMedia  # noqa: F401
+from .rating import Rating, RatingCategory  # noqa: F401
+from .subtitle import Language, Subtitle  # noqa: F401
+from .utils import CODECS  # noqa: F401
+from .utils import ENCODE_EXTENSIONS  # noqa: F401
+from .utils import ENCODE_EXTENSIONS_KEYS  # noqa: F401
+from .utils import ENCODE_RESOLUTIONS  # noqa: F401
+from .utils import ENCODE_RESOLUTIONS_KEYS  # noqa: F401
+from .utils import MEDIA_ENCODING_STATUS  # noqa: F401
+from .utils import MEDIA_STATES  # noqa: F401
+from .utils import MEDIA_TYPES_SUPPORTED  # noqa: F401
+from .utils import category_thumb_path  # noqa: F401
+from .utils import encoding_media_file_path  # noqa: F401
+from .utils import generate_uid  # noqa: F401
+from .utils import original_media_file_path  # noqa: F401
+from .utils import original_thumbnail_file_path  # noqa: F401
+from .utils import subtitles_file_path  # noqa: F401
+from .utils import validate_rating  # noqa: F401
+from .video_data import VideoChapterData, VideoTrimRequest  # noqa: F401
