+++ templates/cms/categories.html
@@ -41,5 +41,5 @@
 {% block content %}<div id="page-categories"></div>{% endblock %}
 
 {% block bottomimports %}
-<script src="{% static "js/categories.js" %}"></script>
+<script src="{% static "js/categories.js" %}?v={{ VERSION }}"></script>
 {% endblock bottomimports %}
