+++ templates/cms/embed.html
@@ -9,5 +9,5 @@
 {% block content %}<div id="page-embed"></div>{% endblock content %}
 
 {% block bottomimports %}
-<script src="{% static "js/embed.js" %}"></script>
+    <script src="{% static "js/embed.js" %}?v={{ VERSION }}"></script>
 {% endblock bottomimports %}
