+++ templates/cms/featured-media.html
@@ -43,5 +43,5 @@
 {% endblock %}
 
 {% block bottomimports %}
-<script src="{% static "js/featured.js" %}"></script>
+    <script src="{% static "js/featured.js" %}?v={{ VERSION }}"></script>
 {% endblock bottomimports %}
