+++ templates/cms/history.html
@@ -13,5 +13,5 @@
 {% endblock %}
 
 {% block bottomimports %}
-<script src="{% static "js/history.js" %}"></script>
+    <script src="{% static "js/history.js" %}?v={{ VERSION }}"></script>
 {% endblock bottomimports %}
