+++ templates/cms/index.html
@@ -48,5 +48,5 @@
 {% block content %}<div id="page-home"></div>{% endblock %}
 
 {% block bottomimports %}
-<script src="{% static "js/index.js" %}"></script>
+	<script src="{% static "js/index.js" %}?v={{ VERSION }}"></script>
 {% endblock bottomimports %}
