+++ templates/cms/latest-media.html
@@ -43,5 +43,5 @@
 {% endblock %}
 
 {% block bottomimports %}
-<script src="{% static "js/latest.js" %}"></script>
+    <script src="{% static "js/latest.js" %}?v={{ VERSION }}"></script>
 {% endblock bottomimports %}
