+++ templates/cms/liked_media.html
@@ -13,5 +13,5 @@
 {% endblock %}
 
 {% block bottomimports %}
-<script src="{% static "js/liked.js" %}"></script>
+    <script src="{% static "js/liked.js" %}?v={{ VERSION }}"></script>
 {% endblock bottomimports %}
