+++ templates/cms/manage_comments.html
@@ -14,5 +14,5 @@
 {% block content %}<div id="page-manage-comments"></div>{% endblock %}
 
 {% block bottomimports %}
-<script src="{% static "js/manage-comments.js" %}"></script>
+    <script src="{% static "js/manage-comments.js" %}?v={{ VERSION }}"></script>
 {% endblock bottomimports %}
