+++ templates/cms/manage_users.html
@@ -14,5 +14,5 @@
 {% block content %}<div id="page-manage-users"></div>{% endblock %}
 
 {% block bottomimports %}
-<script src="{% static "js/manage-users.js" %}"></script>
+    <script src="{% static "js/manage-users.js" %}?v={{ VERSION }}"></script>
 {% endblock bottomimports %}
