+++ templates/cms/media.html
@@ -129,5 +129,5 @@
 {% block content %}<div id="page-media"></div>{% endblock content %}
 
 {% block bottomimports %}
-<script src="{% static "js/media.js" %}"></script>
+    <script src="{% static "js/media.js" %}?v={{ VERSION }}"></script>
 {% endblock bottomimports %}
