+++ templates/cms/members.html
@@ -41,5 +41,5 @@
 {% block content %}<div id="page-members"></div>{% endblock %}
 
 {% block bottomimports %}
-<script src="{% static "js/members.js" %}"></script>
+    <script src="{% static "js/members.js" %}?v={{ VERSION }}"></script>
 {% endblock bottomimports %}
