+++ templates/cms/playlist.html
@@ -13,5 +13,5 @@
 {% endblock %}
 
 {% block bottomimports %}
-<script src="{% static "js/playlist.js" %}"></script>
+    <script src="{% static "js/playlist.js" %}?v={{ VERSION }}"></script>
 {% endblock bottomimports %}
