+++ templates/cms/recommended-media.html
@@ -43,5 +43,5 @@
 {% endblock %}
 
 {% block bottomimports %}
-<script src="{% static "js/recommended.js" %}"></script>
+    <script src="{% static "js/recommended.js" %}?v={{ VERSION }}"></script>
 {% endblock bottomimports %}
