+++ templates/cms/search.html
@@ -6,5 +6,5 @@
 {% endblock %}
 
 {% block bottomimports %}
-<script src="{% static "js/search.js" %}"></script>
+    <script src="{% static "js/search.js" %}?v={{ VERSION }}"></script>
 {% endblock bottomimports %}
