+++ templates/cms/tags.html
@@ -41,5 +41,5 @@
 {% block content %}<div id="page-tags"></div>{% endblock %}
 
 {% block bottomimports %}
-<script src="{% static "js/tags.js" %}"></script>
+    <script src="{% static "js/tags.js" %}?v={{ VERSION }}"></script>
 {% endblock bottomimports %}
