+++ templates/cms/user.html
@@ -29,5 +29,5 @@ No such user
 {% endblock %}
 
 {% block bottomimports %}
-<script src="{% static "js/profile-media.js" %}"></script>
+    <script src="{% static "js/profile-media.js" %}?v={{ VERSION }}"></script>
 {% endblock bottomimports %}
