+++ templates/cms/user_about.html
@@ -29,5 +29,5 @@ No such user
 {% endblock %}
 
 {% block bottomimports %}
-<script src="{% static "js/profile-about.js" %}"></script>
+    <script src="{% static "js/profile-about.js" %}?v={{ VERSION }}"></script>
 {% endblock bottomimports %}
