+++ templates/cms/user_playlists.html
@@ -29,5 +29,5 @@ No such user
 {% endblock %}
 
 {% block bottomimports %}
-<script src="{% static "js/profile-playlists.js" %}"></script>
+    <script src="{% static "js/profile-playlists.js" %}?v={{ VERSION }}"></script>
 {% endblock bottomimports %}
